Create README.md
This 2 player game is developed using HTML, CSS & JavaScript for resolving an argument in decision making. Each player is given a die, all they have to do is refresh the web page then both the die will roll and a random number will be generated. The die with a larger number will win the game and the respective player can make a decision on the argument. This is a fun game if played with friends and family to decide to-go-to which restaurant for dining.etc.